Fix a big security hole exposing aria2c
I've added

 --rpc-listen-all=false

to disable access to aria2c from over the internet. This is a big security hole as by default this app doesn't support rpc-secret or any kinda aria2c authentication. So at least to mitigate the access to aria2c, we can use the edited argument to disable all incoming access from over internet but other than localhost.

Also I tested with this option and it's working fine.